requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pandas
aiofiles
